+++ resources/views/components/tabs/tab.blade.php
@@ -1,4 +1,4 @@
-@props([
+@props ([
     'tabId',
     'tabLabel' => null,
 ])