+++ src/Assets/App/resources/_partials/file-edit-site-toggle.blade.php
@@ -1,8 +1,10 @@
 @if (count(\Thinktomorrow\Chief\Sites\ChiefSites::locales()) > 1)
     <x-chief::tabs :show-nav="true" :should-listen-for-external-tab="false" reference="file-edit-site-toggle">
         @foreach (\Thinktomorrow\Chief\Sites\ChiefSites::locales() as $locale)
-            <x-chief::tabs.tab tab-id="{{ $locale }}"
-                               tab-label="{{ \Thinktomorrow\Chief\Sites\ChiefSites::name($locale) }}"></x-chief::tabs.tab>
+            <x-chief::tabs.tab
+                tab-id="{{ $locale }}"
+                tab-label="{{ \Thinktomorrow\Chief\Sites\ChiefSites::name($locale) }}"
+            />
         @endforeach
     </x-chief::tabs>
 @endif